+++ jforgame-server/src/main/java/com/kingston/jforgame/server/game/activity/ActivityDataPool.java
@@ -3,7 +3,7 @@ package com.kingston.jforgame.server.game.activity;
 public class ActivityDataPool {
 	
 	/** 请求-打开面板 */
-	public static final int REQ_OPEN_PANEL = 1;
+	public static final byte REQ_OPEN_PANEL = 1;
 	
 
 }
