+++ jforgame-server/src/main/java/com/kingston/jforgame/server/game/cross/ladder/service/LadderDataPool.java
@@ -4,12 +4,12 @@ public class LadderDataPool {
 	
 	
 	/** 游戏服--请求报名 */
-	public static final int REQ_LADDER_APPLY = 1;
+	public static final byte REQ_LADDER_APPLY = 1;
 	/** 游戏服--推送传输开始 */
-	public static final int RES_LADDER_TRANSFER_BEGIN = 2;
+	public static final byte RES_LADDER_TRANSFER_BEGIN = 2;
 	/** 战斗服--请求玩家登录 */
-	public static final int RES_LADDER_LOGIN = 3;
+	public static final byte RES_LADDER_LOGIN = 3;
 	/** 战斗服--请求玩家登录 */
-	public static final int REQ_G2F_TRANSFER = 4;
+	public static final byte REQ_G2F_TRANSFER = 4;
 
 }
