+++ jforgame-server/src/main/java/com/kingston/jforgame/server/game/skill/SkillDataPool.java
@@ -4,6 +4,6 @@ public class SkillDataPool {
 
 	//cmd请求协议枚举
 	/** 请求－释放技能 */
-	public static final int REQ_USE_SKILL = 1;
+	public static final byte REQ_USE_SKILL = 1;
 
 }
